+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Reserved auto-selected PaperTiger HTTP ports until Bandit takes ownership of the listener socket, eliminating an intermittent `:eaddrinuse` startup race on Linux CI.
+
 ## [1.1.1] - 2026-05-21
 
 ### Added